Initial commit of main.py
This is the initial commit of main.py in the handwritten-digit-recognition-tensorflow repository. Currently, the code consists of importing TensorFlow and TensorFlow Datasets libraries and loading the MNIST dataset. The 'shuffle_files' parameter is set to 'True' to shuffle the data while training, and 'as_supervised' parameter is set to 'True' to return a tuple of image and label instead of a dictionary. This is the first step in the series of implementations for handwritten digit recognition using TensorFlow.